# creature-lab/api/openapi.yaml
openapi: 3.0.3
info:
  title: Creature Lab API
  description: |
    The alien species catalog for mars-sim, with an SVG map sprite for every
    form of each species' life. See docs/creature-lab.md.

    Every operation needs a credential unless it says `security: []`: a `cl_`
    api key or `cl_at_` access token as `Authorization: Bearer ...`, or the
    web session cookie. Listing and viewing species and their sprite
    candidates (accepted or not) is public, so anyone can browse the catalog. A missing or unknown credential gets a 401 whose
    WWW-Authenticate header points at OAuth discovery.

    This spec is the HTTP surface's source of truth. `make generate` in
    creature-lab/ turns it into internal/creaturelab/api/gen.
  version: 0.2.0
servers:
  - url: /api
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /species:
    get:
      operationId: listSpecies
      summary: List saved species, newest first
      security: []
      responses:
        "200":
          description: The catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeciesList"
    post:
      operationId: createSpecies
      summary: Roll and save a species
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeciesCreate"
      responses:
        "201":
          description: Species created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeciesDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /species/{speciesId}:
    parameters:
      - $ref: "#/components/parameters/SpeciesId"
    get:
      operationId: getSpecies
      summary: One species with its sprite slots
      security: []
      responses:
        "200":
          description: The species
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeciesDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateSpecies
      summary: Replace a species' notes or field notes
      description: Sets whichever fields are given and leaves the rest alone. An empty description resets the field notes to the generated text.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeciesUpdate"
      responses:
        "200":
          description: The updated species
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeciesDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/Unprocessable"
    delete:
      operationId: deleteSpecies
      summary: Remove a species and its sprites
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A cl_ api key, or a cl_at_ OAuth access token.
    cookieAuth:
      type: apiKey
      in: cookie
      name: cl_session
      description: The web session set by /login.
  parameters:
    SpeciesId:
      name: speciesId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: No credential, or one that is unknown, expired or revoked
      content:
        text/plain:
          schema:
            type: string
    NotFound:
      description: No such species or candidate
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unprocessable:
      description: No such form, or an SVG that was refused
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    OK:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    Traits:
      type: object
      description: A species in words. The stored row keeps the sim.AlienSpecies itself; this is its presentation.
      required: [singular, plural, scientificName, temperament, skin, color, pattern, eyes, arms, legs, tail, wings, heightCM, weightKG, attacks, description]
      properties:
        singular:
          type: string
        plural:
          type: string
        scientificName:
          type: string
        emoji:
          type: string
        temperament:
          type: string
        apex:
          type: boolean
          description: One of the rare, very deadly species.
        skin:
          type: string
        color:
          type: string
        pattern:
          type: string
        eyes:
          type: integer
        arms:
          type: integer
        legs:
          type: integer
        tail:
          type: boolean
        wings:
          type: boolean
        heightCM:
          $ref: "#/components/schemas/Range"
        weightKG:
          $ref: "#/components/schemas/Range"
        attacks:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
        description:
          type: string
    Range:
      type: array
      description: "[min, max]"
      minItems: 2
      maxItems: 2
      items:
        type: integer
    Form:
      type: object
      description: One sprite slot, a stage or caste of the species' life. index matches sim.AlienSpecies.LifeForms().
      required: [index, name, label, stage, sizePct, arms, legs]
      properties:
        index:
          type: integer
        name:
          type: string
          description: '"egg", "grub", "queen"; "adult" for a plain adult'
        label:
          type: string
          description: 'What the game calls one: "grelk egg", "grelk"'
        stage:
          type: integer
        sizePct:
          type: integer
        inert:
          type: boolean
        arms:
          type: integer
        legs:
          type: integer
        tail:
          type: boolean
        wings:
          type: boolean
        features:
          type: array
          items:
            type: string
        lays:
          type: boolean
    FormView:
      description: A slot and where its sprite stands.
      allOf:
        - $ref: "#/components/schemas/Form"
        - type: object
          required: [candidates]
          properties:
            acceptedId:
              type: string
            acceptedSvgUrl:
              type: string
            candidates:
              type: integer
...
    SpeciesSummary:
      type: object
      required: [id, seed, singular, plural, scientificName, temperament, formCount, acceptedCount, complete, createdAt, url]
      properties:
        id:
          type: string
        seed:
          type: integer
          format: int64
        singular:
          type: string
        plural:
          type: string
        scientificName:
          type: string
        emoji:
          type: string
        temperament:
          type: string
        apex:
          type: boolean
        formCount:
          type: integer
        acceptedCount:
          type: integer
        complete:
          type: boolean
        createdAt:
          type: string
          format: date-time
        url:
          type: string
        spriteUrls:
          type: array
          description: The accepted sprites, in form order.
          items:
            type: string
    SpeciesList:
      type: object
      required: [species]
      properties:
        species:
          type: array
          items:
            $ref: "#/components/schemas/SpeciesSummary"
    SpeciesDetail:
      type: object
      required: [id, seed, generatorRev, traits, generatedDescription, forms, complete, createdAt, url]
      properties:
        id:
          type: string
        seed:
          type: integer
          format: int64
        generatorRev:
          type: string
          description: The mars-sim commit that rolled it.
        notes:
          type: string
        traits:
          $ref: "#/components/schemas/Traits"
        generatedDescription:
          type: string
          description: The field notes the roster code wrote when it rolled the species. traits.description is these unless descriptionEdited.
        descriptionEdited:
          type: boolean
          description: traits.description was rewritten and differs from generatedDescription.
        forms:
          type: array
          items:
            $ref: "#/components/schemas/FormView"
        complete:
          type: boolean
        createdAt:
          type: string
          format: date-time
        url:
          type: string
    SpeciesCreate:
      type: object
      properties:
        seed:
          type: integer
          format: int64
          description: Seed to roll, usually one from /preview; omit for a random one.
        notes:
          type: string
    SpeciesUpdate:
      type: object
      properties:
        notes:
          type: string
          description: Replacement lab notes; omit to leave them.
          x-go-type-skip-optional-pointer: false
        description:
          type: string
          maxLength: 4000
          description: Replacement field notes; omit to leave them, or send "" to go back to the generated text.
          x-go-type-skip-optional-pointer: false
...